Fixing permissions of files in /etc/tomcat7
By default chown does not traverse symbolic links.  We want to set
the permissions of these files to ensure they are not owned by a
random user.

[Docker documentation](http://docs.docker.io/en/latest/reference/builder/#add)
for ADD claims that the uid/gid of files should both be 0 while the
umask should be 0755.  In testing this has not been found to be the
case.